requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
tqdm>=4.64.0
ipycanvas>=0.13.0
ipywidgets>=8.0.0
scikit-image>=0.19.0
Pillow>=9.0.0
numpy>=1.22.0